requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
langchain-groq>=0.3.2
langchain-community>=0.3.24
langchain-core>=0.3.59,<1.0.0
langgraph>=0.0.19
pydantic>=2.0.0
